+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Changed
+
+- workflow_executor now defaults to the lattice-default executor
+
 ## [0.236.0-rc.0] - 2025-01-21
 
 ### Authors